ADD light classes; REFACTOR aliased some glm types
Added classes for area and point lights as well as a base class for
lights.
Aliased glm types dvec3, dvec2 for easier use and readability.

+++ LiVHY/src/viewport.cpp
@@ -3,7 +3,7 @@ Viewport::Viewport()
 {
 }
 
-Viewport::Viewport(int w, int h, float fov) : m_Width(w), m_Height(h), m_HorizFOV(fov)
+Viewport::Viewport(int w, int h, double fov) : m_Width(w), m_Height(h), m_HorizFOV(fov)
 {
 	CalculateFOV();
 }
@@ -17,11 +17,11 @@ const int Viewport::Height() const
 	return m_Height;
 }
 
-const float Viewport::HFOV() const
+const double Viewport::HFOV() const
 {
 	return m_Width;
 }
-const float Viewport::VFOV() const
+const double Viewport::VFOV() const
 {
 	return m_Height;
 }
@@ -32,12 +32,12 @@ void Viewport::Resize(int w, int h)
 	m_Height = h;
 }
 
-void Viewport::SetFOV(float hfov){
+void Viewport::SetFOV(double hfov){
 	m_HorizFOV = hfov;
 	CalculateFOV();
 }
 
 void Viewport::CalculateFOV(){
 	
-	m_VertFOV = 2.0f * atanf(tanf(m_HorizFOV/2.0f) * m_Height/m_Width);
+	m_VertFOV = 2.0 * atan(tan(m_HorizFOV/2.0) * m_Height/m_Width);
 }
